data-pipeline/requirements.txt
osmium>=3.6.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0
requests>=2.31.0
shapely>=2.0.0